contact.js moved to components folder and reference updated in about.js
